requirements.txt
datasets==2.18.0
matplotlib==3.8.3
numpy==1.26.4
packaging==24.0
Pillow==10.2.0
pycocotools==2.0.7
PyYAML==6.0.1
scipy==1.12.0
termcolor==2.4.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.2
wandb==0.16.5
diffusers==0.31.0
accelerate==1.2.0
transformers==4.47.0
opencv-python==4.10.0.84
inflect==7.4.0